Move permission utilities from validate_permissions.py to utils.py
Consolidate permission checking functions into workflows/utils.py
since they are utilities, not workflows. Remove the separate
validate_permissions.py file.
